requirements.txt
# python 3.11.8
faster-whisper==1.0.2
ffmpy==0.3.2
gradio==4.36.0
gradio_client
pandas==2.0.3
pydub==0.25.1
openpyxl
python-docx==1.1.2
edge-tts==6.1.12
numpy==1.26.4
torch
